requirements.txt
boto3>=1.12,<2.0
cryptography>=3.0,<4.0
nest_asyncio>=1.4,<2.0
pycryptodome>=3.9,<4.0
python-jose>=3.0,<4.0
requests>=2.23,<3.0
requests_toolbelt>=0.9.0,<1.0
six>=1.14,<2.0
websockets>=10.2
pydantic>=1.6,<2.0
tabulate>=0.8.7,<1.0
tqdm>=4.42.1,<5.0
